Use SharpCompress on all platforms now that .tar.gz is needed for Windows Tor bundle
Update dependencies
