+++ packages.config
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="xunit.runner.console" version="2.4.1" />
+  <package id="xunit.runner.console" version="2.4.2" />
 </packages>
